+++ src/syscalls/write.c
@@ -4,7 +4,7 @@
 #include <sched.h>
 
 isize sys_write(const trapframe_t *frame) {
-    isize fd = frame->a0;
+    i32 fd = frame->a0;
     const u8 *buf = (const u8 *) frame->a1;
     usize size = frame->a2;
     if (is_bad_fd(fd))
